requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
groq==0.11.0
duckduckgo-search==5.3.0
python-dotenv==1.0.1
slowapi==0.1.9
requests==2.31.0